Rename changePassword constants file to Authentication
