+++ src/app/constants/Authentication.js
