+++ src/app/views/new-password/changePasswordConfirmed.jsx
@@ -1,10 +1,16 @@
 import React from "react";
 import Panel from "../../components/panel/Panel";
+import PropTypes from "prop-types";
+
+const propTypes = {
+    handleClick: PropTypes.func,
+};
 
 const ChangePasswordConfirmed = props => {
     const changePasswordPanelBody = [
         <p key="info-technical">
-            If you have technical problems with your account, please email&nbsp;<a href="mailto:publishing@ons.gov.uk">publishing@ons.gov.uk</a>
+            If you have technical problems with your account, please email&nbsp;<a
+            href="mailto:publishing@ons.gov.uk">publishing@ons.gov.uk</a>
         </p>
     ];
 
@@ -12,13 +18,16 @@ const ChangePasswordConfirmed = props => {
         <div className="grid grid--justify-center">
             <div className="grid__col-3">
                 <h1>Your password has been changed</h1>
-                <button type="button" onClick={props.handleClick} className="btn btn--primary margin-bottom--2 margin-right--5">
+                <button type="button" onClick={props.handleClick}
+                        className="btn btn--primary margin-bottom--2 margin-right--5">
                     Start using Florence
                 </button>
-                <Panel type={"information"} body={changePasswordPanelBody} />
+                <Panel type={"information"} body={changePasswordPanelBody}/>
             </div>
         </div>
     );
 };
+
+ChangePasswordConfirmed.propTypes = propTypes;
 
 export default ChangePasswordConfirmed;
